Speech/requirementsSpeech.txt
# speech to text
speechrecognition

# openai chat response
python-dotenv
openai

# text to speech
pyttsx3
pygame